requirements.txt
langchain 
langchain-community
langchain-openai
langchainhub 
openai
tiktoken
azure-ai-documentintelligence
requests
azure-search-documents
azure-storage-blob
tenacity
tiktoken
